refactor: switch header to ssr
